모듈 시스템 확장 중, 창고모듈 등 구현중 todo:현행 완료

+++ src/main/java/com/yourname/mycreateaddon/content/kinetics/node/OreNodeBlockEntity.java
@@ -13,7 +13,6 @@ import net.minecraft.nbt.ListTag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
@@ -100,7 +99,7 @@ public class OreNodeBlockEntity extends SmartBlockEntity implements IHaveGoggleI
         }
 
         float effectiveMiningAmount = miningAmount / getHardness();
-        this.miningProgress += effectiveMiningAmount;
+        this.miningProgress += (int) effectiveMiningAmount;
 
         int miningResistance = 1000;
